+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/websocket_api/models/enums.py
@@ -96,11 +96,6 @@ class NewOrderTimeInForceEnum(Enum):
     RPI = "RPI"
 
 
-class NewOrderWorkingTypeEnum(Enum):
-    MARK_PRICE = "MARK_PRICE"
-    CONTRACT_PRICE = "CONTRACT_PRICE"
-
-
 class NewOrderNewOrderRespTypeEnum(Enum):
     ACK = "ACK"
     RESULT = "RESULT"