+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/websocket_api/websocket_api.py
@@ -58,7 +58,6 @@ from .models import NewAlgoOrderSelfTradePreventionModeEnum
 from .models import NewOrderSideEnum
 from .models import NewOrderPositionSideEnum
 from .models import NewOrderTimeInForceEnum
-from .models import NewOrderWorkingTypeEnum
 from .models import NewOrderNewOrderRespTypeEnum
 from .models import NewOrderPriceMatchEnum
 from .models import NewOrderSelfTradePreventionModeEnum
@@ -449,7 +448,7 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
         symbol: Union[str, None],
         side: Union[ModifyOrderSideEnum, None],
         quantity: Union[float, None],
-        price: Optional[float] = None,
+        price: Union[float, None],
         id: Optional[str] = None,
         order_id: Optional[int] = None,
         orig_client_order_id: Optional[str] = None,
@@ -477,11 +476,11 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
                     symbol (Union[str, None]):
                     side (Union[ModifyOrderSideEnum, None]): `SELL`, `BUY`
                     quantity (Union[float, None]): Order quantity, cannot be sent with `closePosition=true`
-                    price (Optional[float] = None):
+                    price (Union[float, None]):
                     id (Optional[str] = None): Unique WebSocket request ID.
                     order_id (Optional[int] = None):
                     orig_client_order_id (Optional[str] = None):
-                    price_match (Optional[ModifyOrderPriceMatchEnum] = None): only avaliable for `LIMIT`/`STOP`/`TAKE_PROFIT` order; can be set to `OPPONENT`/ `OPPONENT_5`/ `OPPONENT_10`/ `OPPONENT_20`: /`QUEUE`/ `QUEUE_5`/ `QUEUE_10`/ `QUEUE_20`; Can't be passed together with `price`
+                    price_match (Optional[ModifyOrderPriceMatchEnum] = None): only available for `LIMIT` order; can be set to `OPPONENT`/ `OPPONENT_5`/ `OPPONENT_10`/ `OPPONENT_20`/ `QUEUE`/ `QUEUE_5`/ `QUEUE_10`/ `QUEUE_20`; Can't be passed together with `price`
                     recv_window (Optional[int] = None):
 
                 Returns:
@@ -576,14 +575,14 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
                     id (Optional[str] = None): Unique WebSocket request ID.
                     position_side (Optional[NewAlgoOrderPositionSideEnum] = None): Default `BOTH` for One-way Mode ; `LONG` or `SHORT` for Hedge Mode. It must be sent in Hedge Mode.
                     time_in_force (Optional[NewAlgoOrderTimeInForceEnum] = None):
-                    quantity (Optional[float] = None): Cannot be sent with `closePosition`=`true`(Close-All)
+                    quantity (Optional[float] = None):
                     price (Optional[float] = None):
                     trigger_price (Optional[float] = None):
-                    working_type (Optional[NewAlgoOrderWorkingTypeEnum] = None): stopPrice triggered by: "MARK_PRICE", "CONTRACT_PRICE". Default "CONTRACT_PRICE"
-                    price_match (Optional[NewAlgoOrderPriceMatchEnum] = None): only avaliable for `LIMIT`/`STOP`/`TAKE_PROFIT` order; can be set to `OPPONENT`/ `OPPONENT_5`/ `OPPONENT_10`/ `OPPONENT_20`: /`QUEUE`/ `QUEUE_5`/ `QUEUE_10`/ `QUEUE_20`; Can't be passed together with `price`
-                    close_position (Optional[str] = None): `true`, `false`；Close-All，used with `STOP_MARKET` or `TAKE_PROFIT_MARKET`.
-                    price_protect (Optional[str] = None): "true" or "false", default "false". Used with `STOP/STOP_MARKET` or `TAKE_PROFIT/TAKE_PROFIT_MARKET` orders.
-                    reduce_only (Optional[str] = None): "true" or "false". default "false". Cannot be sent in Hedge Mode; cannot be sent with `closePosition`=`true`
+                    working_type (Optional[NewAlgoOrderWorkingTypeEnum] = None): triggerPrice triggered by: `MARK_PRICE`, `CONTRACT_PRICE`. Default `CONTRACT_PRICE`
+                    price_match (Optional[NewAlgoOrderPriceMatchEnum] = None): only available for `LIMIT` order; can be set to `OPPONENT`/ `OPPONENT_5`/ `OPPONENT_10`/ `OPPONENT_20`/ `QUEUE`/ `QUEUE_5`/ `QUEUE_10`/ `QUEUE_20`; Can't be passed together with `price`
+                    close_position (Optional[str] = None): true, false；Close-All，used with `STOP_MARKET` or `TAKE_PROFIT_MARKET`.
+                    price_protect (Optional[str] = None): "true" or "false", default "false". Used with `STOP_MARKET` or `TAKE_PROFIT_MARKET` order. when price reaches the triggerPrice ，the difference rate between "MARK_PRICE" and "CONTRACT_PRICE" cannot be larger than the Price Protection Threshold of the symbol.
+                    reduce_only (Optional[str] = None): "true" or "false". default "false". Cannot be sent in Hedge Mode
                     activate_price (Optional[float] = None): Used with `TRAILING_STOP_MARKET` orders, default as the latest price(supporting different `workingType`)
                     callback_rate (Optional[float] = None): Used with `TRAILING_STOP_MARKET` orders, min 0.1, max 10 where 1 for 1%
                     client_algo_id (Optional[str] = None):
@@ -637,12 +636,6 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
         reduce_only: Optional[str] = None,
         price: Optional[float] = None,
         new_client_order_id: Optional[str] = None,
-        stop_price: Optional[float] = None,
-        close_position: Optional[str] = None,
-        activation_price: Optional[float] = None,
-        callback_rate: Optional[float] = None,
-        working_type: Optional[NewOrderWorkingTypeEnum] = None,
-        price_protect: Optional[str] = None,
         new_order_resp_type: Optional[NewOrderNewOrderRespTypeEnum] = None,
         price_match: Optional[NewOrderPriceMatchEnum] = None,
         self_trade_prevention_mode: Optional[
@@ -656,40 +649,12 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
 
                 Send in a new order.
 
-        * Order with type `STOP`,  parameter `timeInForce` can be sent ( default `GTC`).
-        * Order with type `TAKE_PROFIT`,  parameter `timeInForce` can be sent ( default `GTC`).
-        * Condition orders will be triggered when:
-
-        * If parameter`priceProtect`is sent as true:
-        * when price reaches the `stopPrice` ，the difference rate between "MARK_PRICE" and "CONTRACT_PRICE" cannot be larger than the "triggerProtect" of the symbol
-        * "triggerProtect" of a symbol can be got from `GET /fapi/v1/exchangeInfo`
-
-        * `STOP`, `STOP_MARKET`:
-        * BUY: latest price ("MARK_PRICE" or "CONTRACT_PRICE") >= `stopPrice`
-        * SELL: latest price ("MARK_PRICE" or "CONTRACT_PRICE") <= `stopPrice`
-        * `TAKE_PROFIT`, `TAKE_PROFIT_MARKET`:
-        * BUY: latest price ("MARK_PRICE" or "CONTRACT_PRICE") <= `stopPrice`
-        * SELL: latest price ("MARK_PRICE" or "CONTRACT_PRICE") >= `stopPrice`
-        * `TRAILING_STOP_MARKET`:
-        * BUY: the lowest price after order placed `<= `activationPrice`, and the latest price >`= the lowest price * (1 + `callbackRate`)
-        * SELL: the highest price after order placed >= `activationPrice`, and the latest price <= the highest price * (1 - `callbackRate`)
-
-        * For `TRAILING_STOP_MARKET`, if you got such error code.
-        ``{"code": -2021, "msg": "Order would immediately trigger."}``
-        means that the parameters you send do not meet the following requirements:
-        * BUY: `activationPrice` should be smaller than latest price.
-        * SELL: `activationPrice` should be larger than latest price.
-
         * If `newOrderRespType ` is sent as `RESULT` :
         * `MARKET` order: the final FILLED result of the order will be return directly.
         * `LIMIT` order with special `timeInForce`: the final status result of the order(FILLED or EXPIRED) will be returned directly.
 
-        * `STOP_MARKET`, `TAKE_PROFIT_MARKET` with `closePosition`=`true`:
-        * Follow the same rules for condition orders.
-        * If triggered，**close all** current long position( if `SELL`) or current short position( if `BUY`).
-        * Cannot be used with `quantity` paremeter
-        * Cannot be used with `reduceOnly` parameter
-        * In Hedge Mode,cannot be used with `BUY` orders in `LONG` position side. and cannot be used with `SELL` orders in `SHORT` position side
+        * `selfTradePreventionMode` is only effective when `timeInForce` set to `IOC` or `GTC` or `GTD`.
+        * In extreme market conditions, timeInForce `GTD` order auto cancel time might be delayed comparing to `goodTillDate`
 
         Weight: 0
 
@@ -700,18 +665,12 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
                     id (Optional[str] = None): Unique WebSocket request ID.
                     position_side (Optional[NewOrderPositionSideEnum] = None): Default `BOTH` for One-way Mode ; `LONG` or `SHORT` for Hedge Mode. It must be sent in Hedge Mode.
                     time_in_force (Optional[NewOrderTimeInForceEnum] = None):
-                    quantity (Optional[float] = None): Cannot be sent with `closePosition`=`true`(Close-All)
-                    reduce_only (Optional[str] = None): "true" or "false". default "false". Cannot be sent in Hedge Mode; cannot be sent with `closePosition`=`true`
+                    quantity (Optional[float] = None):
+                    reduce_only (Optional[str] = None): "true" or "false". default "false". Cannot be sent in Hedge Mode
                     price (Optional[float] = None):
                     new_client_order_id (Optional[str] = None): A unique id among open orders. Automatically generated if not sent. Can only be string following the rule: `^[.A-Z:/a-z0-9_-]{1,36}$`
-                    stop_price (Optional[float] = None): Used with `STOP/STOP_MARKET` or `TAKE_PROFIT/TAKE_PROFIT_MARKET` orders.
-                    close_position (Optional[str] = None): `true`, `false`；Close-All，used with `STOP_MARKET` or `TAKE_PROFIT_MARKET`.
-                    activation_price (Optional[float] = None): Used with `TRAILING_STOP_MARKET` orders, default as the latest price(supporting different `workingType`)
-                    callback_rate (Optional[float] = None): Used with `TRAILING_STOP_MARKET` orders, min 0.1, max 10 where 1 for 1%
-                    working_type (Optional[NewOrderWorkingTypeEnum] = None): stopPrice triggered by: "MARK_PRICE", "CONTRACT_PRICE". Default "CONTRACT_PRICE"
-                    price_protect (Optional[str] = None): "true" or "false", default "false". Used with `STOP/STOP_MARKET` or `TAKE_PROFIT/TAKE_PROFIT_MARKET` orders.
                     new_order_resp_type (Optional[NewOrderNewOrderRespTypeEnum] = None): "ACK", "RESULT", default "ACK"
-                    price_match (Optional[NewOrderPriceMatchEnum] = None): only avaliable for `LIMIT`/`STOP`/`TAKE_PROFIT` order; can be set to `OPPONENT`/ `OPPONENT_5`/ `OPPONENT_10`/ `OPPONENT_20`: /`QUEUE`/ `QUEUE_5`/ `QUEUE_10`/ `QUEUE_20`; Can't be passed together with `price`
+                    price_match (Optional[NewOrderPriceMatchEnum] = None): only available for `LIMIT` order; can be set to `OPPONENT`/ `OPPONENT_5`/ `OPPONENT_10`/ `OPPONENT_20`/ `QUEUE`/ `QUEUE_5`/ `QUEUE_10`/ `QUEUE_20`; Can't be passed together with `price`
                     self_trade_prevention_mode (Optional[NewOrderSelfTradePreventionModeEnum] = None): `EXPIRE_TAKER`:expire taker order when STP triggers/ `EXPIRE_MAKER`:expire taker order when STP triggers/ `EXPIRE_BOTH`:expire both orders when STP triggers; default `NONE`
                     good_till_date (Optional[int] = None): order cancel time for timeInForce `GTD`, mandatory when `timeInforce` set to `GTD`; order the timestamp only retains second-level precision, ms part will be ignored; The goodTillDate timestamp must be greater than the current time plus 600 seconds and smaller than 253402300799000
                     recv_window (Optional[int] = None):
@@ -735,12 +694,6 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
             reduce_only,
             price,
             new_client_order_id,
-            stop_price,
-            close_position,
-            activation_price,
-            callback_rate,
-            working_type,
-            price_protect,
             new_order_resp_type,
             price_match,
             self_trade_prevention_mode,