+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/websocket_streams/streams/market_api.py
@@ -682,9 +682,15 @@ class MarketApi:
             /tradingSession
             https://developers.binance.com/docs/derivatives/usds-margined-futures/websocket-market-streams/Trading-Session-Stream
 
-            Trading session information for the underlying assets of TradFi Perpetual contracts—covering the U.S. equity market and the commodity market—is updated every second. Trading session information for different underlying markets is pushed in separate messages. Session types for the equity market include "PRE_MARKET", "REGULAR", "AFTER_MARKET", "OVERNIGHT", and "NO_TRADING". Session types for the commodity market include "REGULAR" and "NO_TRADING".
+            Trading session information for the underlying assets of TradFi Perpetual contracts, covering the U.S. equity market, Korean equity market, and the commodity market, is updated every second. Trading session information for different underlying markets is pushed in separate messages.
 
-        Update Speed: 1s
+            **Event type:**
+
+            - `EquityUpdate`: Session types for the U.S. equity market include "PRE_MARKET", "REGULAR", "AFTER_MARKET", "OVERNIGHT", and "NO_TRADING".
+            - `CommodityUpdate`: Session types for the commodity market include "REGULAR" and "NO_TRADING".
+            - `KR_EquityUpdate`: Session types for the Korean equity market include "REGULAR" and "NO_TRADING".
+
+            Update Speed: 1s
 
             Args:
                     id (Optional[str] = None): Unique WebSocket request ID.